Make node its own component
